+++ app/src/Layout/Layout.css
@@ -1,0 +1,57 @@
+/* Header css */
+.header {
+  background-color: #4b2e83;
+  height: 11vh;
+  width: 100%;
+  position: fixed;
+  z-index: 1;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.logo {
+  color: white;
+  font-size: 2.5vw;
+  font-weight: 700;
+  margin-left: 2%;
+  user-select: none;
+  cursor: pointer;
+}
+
+.login {
+  color: white;
+  font-size: 2vw;
+  font-weight: 400;
+  margin-right: 2%;
+  user-select: none;
+  cursor: pointer;
+}
+
+/* Sidebar css */
+.sidebar {
+  background-color: #f0f0f0;
+  height: 89vh;
+  width: 15%;
+  bottom: 0%;
+  position: fixed;
+}
+
+/* Classtab css */
+.leveltab {
+  text-align: center;
+  min-height: 9%;
+  width: 100%;
+  display: inline-block;
+  border: none;
+  margin-top: 4%;
+  font-size: 140%;
+}
+
+.leveltab:hover,
+.leveltab.clicked {
+  background-color: rgba(255, 255, 255, 0.897);
+  font-weight: bold;
+  background: linear-gradient(to right, #f0f0f0 7%, 7%,
+                            #c985ff 17%, transparent 17%);
+}
